start manager tests, clean up main.go, create game_play for random game logic
